BIG transition clean-up!
The goal of this update is to streamline and actually fix the
transitions to make them work for all of the element types. This was
actually a PITA since the initial implementation, which AI did a decent
job with, was a bit overly complicated.

This clean-up appears to work for all elements but I feel there's more
clean-up we can do to reduce complexity and possibly stack depth.

This also adds proper FPS handling for animated elements.