Fix first send issue from the Jetson to the ESP32
There appears to be garbage in the send buffer when transmitting over
serial usb to the ESP32. I tracked this down to the Jetson-side buffer.
I don't have a fix but just doing a dummy send is a decent work-around.
This only affects the very first send on boot.

I also learned that the MQTT buffer is not guaranteed to be null
terminated if text since that buffer could be just data. So I've fixed
that potential issue as well.